More splitting for settings
